UI improvements and code refactor into smaller components; get GMaps working
