+++ android/app/src/main/AndroidManifest.xml
@@ -18,6 +18,7 @@
       android:theme="@style/AppTheme"
       android:requestLegacyExternalStorage="true"
       android:networkSecurityConfig="@xml/network_security_config">
+      <meta-data android:name="com.google.android.geo.API_KEY" android:value="@string/GMAPS_API_KEY" />
       <activity
         android:name=".MainActivity"
         android:label="@string/app_name"
