+++ env.example
@@ -14,3 +14,5 @@ OAUTH_CLIENT_SECRET=your-client-secret
 # Credentials to log in a user for e2e testing
 E2E_TEST_USERNAME=test-username
 E2E_TEST_PASSWORD=test-password
+
+GMAPS_API_KEY=some-key
