+++ index.js
@@ -19,11 +19,14 @@ import { AppRegistry } from "react-native";
 import Config from "react-native-config";
 import { setJSExceptionHandler, setNativeExceptionHandler } from "react-native-exception-handler";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { enableLatestRenderer } from "react-native-maps";
 import { startNetworkLogging } from "react-native-network-logger";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 
 import { name as appName } from "./app.json";
 import { log } from "./react-native-logs.config";
+
+enableLatestRenderer( );
 
 const logger = log.extend( "index.js" );
