+++ src/components/SharedComponents/Buttons/CloseButton.js
@@ -8,18 +8,23 @@ import { IconButton, useTheme } from "react-native-paper";
 
 type Props = {
   className?: string,
-  handleClose?: Function
+  handleClose?: Function,
+  black?: boolean,
+  size?: number
 }
 
-const CloseButton = ( { className, handleClose }: Props ): Node => {
+const CloseButton = ( {
+  className, handleClose, black, size
+}: Props ): Node => {
   const navigation = useNavigation( );
   const theme = useTheme( );
 
   return (
     <IconButton
       icon="close"
+      size={size}
       className={className}
-      iconColor={theme.colors.background}
+      iconColor={black ? theme.colors.tertiary : theme.colors.background}
       onPress={( ) => {
         if ( handleClose ) {
           handleClose( );
